[kitsune] Move global reachability, cloning and embedded module preparation
A class has been created to compute the set of GlobalValue's reachable from
certain entities. This is to allow the analysis to be more sophisticated in the
future - for instance, to attempt to track indirect uses of globals.

The code to clone these reachable globals was moved into kitsune/lib. The
interface to the code was simplified. The embedded module preparation pass for
Hip was modified to set the visibility and linkage of global variables and
kernel functions. This was moved out of loop spawning. The associated tests were
also modified, and in some cases, improved.

The KitsuneUtils.{h,cpp} file was renamed KitsuneLoweringUtils to better reflect
the purpose of the code contained within it.