[kitsune] Add a pass to create globals for the embedded bitcode and fat binary
The pass has been added to Kitsune's pre-tapir pipeline. This creates an
embedded module if required and populates it with functions containing the
kit_device attribute, as well as all other global values reachable from the
function. For the verifier to succeed, a corresponding global for the fat binary
is also created.

The cuda and hip tapir targets were updated to use the global variables created
by this pass. The code in these tapir targets was also cleaned up.