[kitsune] Move some globals to deal with hip fat binaries into the runtime
Rather than generate a fat binary bundle and handle in the global ctor code,
add these as global variables in the runtime with a fixed name. Provide an
entry point to register the fat binary that is called from the ctor. This
registers an atexit handler to unregister the fat binary, obviating the need to
generate a dtor altogether for hip. This does result in __kithip_destroy not
being called - but that was disabled anyway pending the fix for a bug in the
runtime. Since the runtime itself is currently undergoing a major refactor, this
untidiness is temporarily tolerable.

kithip wrapper functions have been provide to register fat binaries and global
variables. The direct calls to functions from AMD's runtime have been removed.

The runtime now expects that fat binary to be present in a global variable with
a well-known name. This is in preparation for full separate compilation support
where this symbol will be introduced into an executable at link-time.