[kitsune] Reconcile upstream LLVM post merge with 21.x
Some changes from LLVM did not get reflected in the code - other code that was
removed from upstream was retained. All of this has been removed. Some slight
reorganization of Kitsune-specific code was carried out, almost all of it in
clang. This simply consolidated the Kitsune-specific additions into large
blocks. Where it was reasonable to do so, these blocks were moved to the end of
the file, or near the end of the block construct in which they were contained.
This was done to make future merges easier since it reduces the likelihood of
conflicts arising. When conflicts do occur, it ought to be easier to see how to
resolve them since they should occur only at the boundaries of these blocks.