[kitsune] Add require directives to OpenCilk tests
The opencilk tests use -tapir-target=opencilk. It is possible that the opencilk
tapir target is disabled when building Kitsune. In this case, the tests will
fail because Kitsune will be unable to build an OpenCilk tapir target object.
Add a `REQUIRE: kitsune-opencilk` directive to ensure that the tests are only
run when the opencilk tapir target has been enabled.