[kitsune] Add pthreads tapir target
This is a very straightforward tapir target that is intended to be the parallel
analogue to the serial tapir target. The target has an associated runtime that
simply takes the range of iterations in a tapir loop and splits it evenly across
a number of pthreads. This number is determined by the number of cores present
on the system and the number of iterations of the loop. The number of threads
can also be fixed by setting the KIT_NUM_THREADS environment variable.

For this, a new "basic" TapirSpawnStrategy was added. The spawn strategies have
also been renumbered which results in a large number of tests having to be
modified.

The code in CGLoopInfo that handles tapir loop attributes has been refactored
to consolidate the Kitsune-specific code in its own object. This is in keeping
with the general approach of trying to minimize the footprint of the
Kitsune-specific modifications in upstream LLVM. However, the code has not been
moved into kitsune/.

Some utility functions were added for types nad metadata. Some code in the
ctor generation pass was cleaned up and refactored.