[kitsune] Require tapir.loop.target metadata on all tapir loops
Prior to this, hintsDemandOutlining() was used to ensure that a loop that had
the structure of a tapir loop was not accidentally identified as a tapir loop.
Loops were likely to be misidentified because of the actions of one or more
transformation passes within LLVM. However, the use of hintsDemandOutlining()
to resolve such ambiguities made those parts of the code somewhat inscrutable
since it was not clear why outlining was necessary in several cases. This also
complicated development of new tapir targets since they would have to use a
spawn strategy that required outlining, even if the targets themselves did not
require it. This is a breaking change since it imposes strict requirements on
the frontends to generate this annotation. Any LLVM IR that is to be passed to
opt, for instance, also needs to be updated. There is no way for the middle-end
to know that some loop is missing a tapir.loop.target annotation, so there is no
way to raise an error, or issue a warning if loops are encountered that are
missing the annotation - it is too easy to raise false positives.
hintsDemandOutlining() has not been removed from the code since it may yet
prove useful.