[kitsune] Fix tests to generate embedded bitcode modules on the fly
Prior to this change, a large number of tests consisted of embedded LLVM
bitcode modules in the test. These consisted of lines generally greater than
1.5K characters that were base64 encoded. This made them only marginally better
than having binary blobs.

This can cause problems when upgrading to newer versions of LLVM since the
encoded blobs are essentially just the binary LLVM bitcode format for which
neither forward, nor backward, compatibility is guaranteed.

With this change, all such blobs have been removed in favor of generating them
on the fly with the kit-enc utility (which was used to generate them in the
anyway). In some cases, the file consists of the actual contents of the
embedded bitcode in human-readable LLVM assembly. These files are passed to
kit-enc. This is much more in keeping with the standard LLVM pattern where the
RUN lines in tests pass the containing file to opt. This should make it much
easier maintain these tests.

In the process, some related bugs were found and fixed.