[kitsune] Fix issues that occur when the cuda tapir target is not enabled
Some unit tests had an implicit dependence on the cuda tapir target. These
dependences have been removed by using the pthreads tapir target, which is
guaranteed to be built. The KITSUNE_*_VERSION_* preprocessor symbols that are
used by the cuda and hip tapir targets were previously only set when the
corresponding tapir targets were enabled. This causes issues in the code where
these are used for conditional compilation (and could potentially also be used
in if constexpr constructs). These are now unconditionally set to 0, so if the
corresponding tapir targets are not built, the symbols will still be available.