[kitsune] Rename examples/ to examples.old/
This is in preparation for the addition of support for external tapir target
plugins. The contents of examples/ (now in examples.old/) are likely to be
obsolete. Some of the contents that are useful could be moved to examples/ or
the kitsune-test-suite. Anything else, including the examples.old directory
itself should eventually be deleted.