[kitsune] Add 'custom' tapir target
This allows a tapir target to be developed external to Kitsune and then loaded
from a dynamic shared object when application code is being compiled. This is
similar in principle to LLVM's pass plugins.

A demo of such a tapir target has been added to the examples/ directory.
Similar to LLVM's examples, these are optionally built when building Kitsune.
Unlike LLVM's examples, though, Kitsune's examples are built by default unless
-DKITSUNE_BUILD_EXAMPLES=OFF is provided at configure time.

Constructing the tapir target options object can now return an error since
the --tapir-plugin= option is required when --tapir=custom. In addition to this,
it is possible for an error to occur when loading the plugin for a number of
reasons. More extensive checking of the options object will be carried out in
a future commit.

Some minor, unrelated cleanup was also carried out.