[kitsune] Validate tapir target options
Depending on the tapir target, some other options are required when running the
tapir/kitsune lowering passes directly with opt. For instance, the cuda
architecture and libdevice bitcode files are needed by the cuda and hip tapir
targets. Without these, one or more of the Kitsune-specific passes may fail.
When using clang/flang, these are determined by clang and the appropriate
options are passed to cc1/fc1. With this, the user is afforded some level of
protection since opt will fail with a reasonable error instead of crashing or,
at best, triggering an assertion.

The validation is not performed when the tapir/kitsune passes are not enabled.
This is to allow specific passes to be run without requiring the user to
provide, for instance, a valid bitcode file. This should hopefully balance the
competing primitives of providing some safety while still allowing power-users
to do "dangerous" things.

In addition to requiring options, the values of the options are also now
checked. This is intended to bring the behavior of opt a bit closer to
clang/flang where such checks are performed.

Tests for the behavior of llc and lld have been added.