[kitsune] Tweak tests and add dependence on offload-bundler
clang does not currently support cuda 13, and it will be some time before it
does. This tweaks some of the tests so they are not dependent on cuda utilities
whose behavior has changed in 13.

These may also depend on clang's offload-bundler, so add an explicit dependence
to ensure that it is built before the tests are run.