[kitsune] Yet another attempt at fixing cuda13 testing and llvm::Error issues
Use -fsyntax-only in the tests that checked that the -x cuda and -x hip options
were handled correctly depending on the frontend (clang vs. kitxx depending on
whether or not --tapir wasp provided). This should resolve the issue of what
to do if the fatbinary executable (in the case of cuda) is not found or takes
a different set of options (since Kitsune generally supports newer versions of
cuda than clang).

Since TTPlugin::load returns an error, and the llvm::Error objects must be
handled explicitly (although the documentation suggests that a weaker condition
is sufficient), add a utility function that "handles" all errors. Somce this
function uses LLVM's handleAllErrors function, it stops LLVM from complaining
about unhandled errors - even if the "handlers" provided to handleAllErrors by
the utility function does absolutely nothing with the error objects.

In addition, the compiler option used in TTPluginDemo was updated to something
that is present in both clang and flang and is also accepted by the two
drivers as well as cc1 and fc1.

Hopefully, this takes care of these once and for all. There are still cuda13
issues that need to be handled, but that is in the long list of things in
Kitsune that need to be handled.