[kitsune][docs] Add initial scaffolding for documentation
Add all the cmake support to generate both HTML documentation and man pages for
Kitsune. Most of this is borrowed liberally from the flang documentation. Most
of the documentation structure is just outlines with little actual content yet.
The old documentation is preserved, but has been put into kitsune/docs/old/
which is intended to be removed when the corresponding documentation is updated.

The command line option references in both the man pages and the HTML
documentation are largely complete. However, documentation for Kitsune-specific
attributes and extensions are not yet generated from the corresponding clang
documentation.

Documentation for Kitsune's Fortran frontend is generated unconditionally, even
though Fortran support is incomplete.