[kitsune] Fixes to build Kitsune on FreeBSD
Only the opencilk tapir target has been tested, as well as those that are always
built. Only linking against libLLVM has been tested.

The build system now passes lists of required libraries to the toolchain via
the config.h header. With this the code in clang/lib/Driver/ToolChain.cpp does
not need to be updated if the required libraries for any of the tapir targets
or the runtime change. It is sufficient to only change the CMakeLists.txt
scripts.