[kitsune] Squelch spurious warning
The sequential spawn strategy does not outline tapir loops. As a result, the
outlineAllTapirLoops() function in LoopSpawningTI.cpp would emit a warning about
a tapir loop being skipped. While this is technically correct, it can be
confusing since it suggests that something "went wrong" though that is not the
case here. A warning is no longer emitted when the sequential spawn strategy is
used.