[kitsune] Fixes to support OSX
The major part of the work involved finding the correct library names for the
various opencilk components. Another sizable chunk of effort went into adding
sysroots to the tests that compile C/C++ code since, on an Xcode-centric setup,
the sysroot is installed in a very non-standard location.

Of the "optional" tapir targets, only opencilk has been tested.