[kitsune] Stop prefetch pass from generating too many prefetches
The tests have been modified to check that exactly the correct number of
prefetches are expected. They were also modified to check that prefetches are
issued for all arrays used in the parallel for loop.

In the future, if the prefetching logic is made more complicated, the tests may
have to be rewritten/split, but for now, they are sufficient.