[kitsune] Add kitsune theme and documentation for Kitsune's tools
Add an Sphinx HTML theme for Kitsune. This is based on the nature theme that
was used earlier, but with some small tweaks.

Remove some static elements and themes that were copied over from the flang
documentation (that provided the initial scaffolding this documentation), but is
not actually used.

Add documentation for Kitsune-specific command-line tools and utilities.

Reorganize some of the documentation and tweak some small things here and there.
The build system for the documentation was also improved in the process.

Some minor incidental code clean up has also been perfomed, in addition to
tweaking the help messages in some of hte utilities themselves.