[kitsune] Remove --kitsune-config-dir command-line option
In the past, it made sense to have a separate directory for  Kitsune-specific
configuration files because we did not have a dedicated driver and just used
`clang`. Now, we have `kitcc`, `kit++`, and `kitfc` that live  alongside
`clang`. This means that we can have dedicated configuration files  for these
drivers as well as clang and flang in the same directory.

If a dedicated configuration directory is desired (or required), the
`--config-system-dir`  and `--config-user-dir` options are available. These
can also be configured at build-time with the CLANG_CONFIG_FILE_SYSTEM_DIR and
CLANG_CONFIG_FILE_USER_DIR configure-time options if needed.