[kitsune] Remove configs/ directory
The configs/ directory contained "sample" configuration files for the various
tapir targets. These were just empty with a comment describing what they are
intended for and where they should be installed. This was somewhat helpful
since there was no other documentation describing how to use configuration
files. The configuration files themselves were also a lot more important then
since the driver was not up to the task of ensuring that the correct
preprocessor, compiler and linker options were added when compiling with a
given tapir target.

The driver is much more capable now which means that the configuration files are
likely to be needed only in more specific situations. In addition, we will
soon have HTML documentation describing configuration files (for both drivers
and individual tapir targets) in more detail, so there is little point in
keeping these, now largely redundant, files around. Removing them also
simplifies the process of adding a new tapir target by eliminating the need for
some additional steps.