[kitsune][docs] Add support for doxygen-generated API documentation
The doxygen-generated documentation is installed into the doxygen/ subdirectory
within the HTML documentation. A link to this has been added to the main HTML
documentation page. Since Kitsune allows building the HTML documentation without
building the API reference, a page has been added with a message stating that
the API documentation is not available. If API documentation is built, this
will not be installed and the doxygen documentation will be installed instead.

The doxygen landing page contains a link to a page that contains a listing of
Kitsune-specific entities (classes, functions etc.). This required adding these
entities to a dedicated documentation group. All of Kitsune's header files were
modified in order to support this. The implementation (.cpp) files were not
modified, so entities from there will not appear in the listing. Documentation
for those are generated, they just don't appear in the listing.

Some documentation was also tweaked in the process for better readability etc.

A configuration option is available to build LLVM's doxygen-generated
documentation alongside Kitsune's. This makes Kitsune's API documentation more
useful since the LLVM classes, enums etc can be browsed together with Kitsune's.
This must be enabled explicitly at configure-time since building LLVM's API
documentation requires more time and 2G of disk space.