[kitsune][docs] Write documenation on writing a tapir target plugin
In addition to the documentation, the KitsuneConfig.cmake file has been
extensively modified. Until now, this was unused, and therefore, contained
details that were only relevant to a very early design of Kitsune that has
since been rendered obsolete.

This file also exposes a subset of the information provided by LLVMConfig.cmake.
With this, users need only use `find_package(KitsuneConfig)` instead of having
to also use `find_package(LLVM)`. The LLVM-related information that is exported
is the subset that is likely to be useful to users interested in writing plugins
for Kitsune. In the future, more variables may be exported if required.

In addition to this, some Find*.cmake files in kitsune/cmake/modules that are
now obsolete have been deleted. There are more Find*.cmake files - related to
the currently unsupported qthreads and realm tapir targets - that could also be
deleted, but we have not done so here.