[kitsune][docs] Add documentation for Kitsune-specific attributes
These are automatically generated from the documentation in clang/include.

That documentation - and the attribute definitions - were moved into their own
files. This is not strictly necessary, but it is in keeping with our effort to
keep Kitsune-specific code as self-contained and as separate as reasonably
possible from the main LLVM project code.

The existing attribute documentation was also brought up to date and, in some
cases, refined. This is only partially complete - there is quite a bit more that
couldl be done. However, we defer that until later since there is a longer
discussion to be had about the Kitsune "language" and end-user interface, so
there is little point in cleaning up something that may well change in the
not-so-distant future.