[kitsune][docs] Remove link to the limitations document from the landing page
Move the document to the old/ directory. It is not clear how relevant those
limitations are. In any case, they can be integrated into the text in a more
useful way than having a separate limitations page. So we keep the text around
in case it is useful at some point.