[kitsune][docs] Miscellaneous documentation fixes
Moves Kitsune's LLVM attribute definitions into a separate file. The file is
included in the main LLVM attributes tablegen definition file.  This is in
keeping with the general approach of isolating Kitsune-specific code as much as
possible.

Add llvm/Transforms/Tapir/ to the doxygen configuration, so the tapir targets
are included when generating Kitsune's API reference. The appropriate classes
and functions are added to the `kitsune` doxygen group.

Update the doxygen documentation to the latest version.

Fix inline documentation in Kitsune so Doxygen doesn't complain about missing
missing references, undocumented parameters and the like.