[kitsune][docs] Add skeleton documents for LLVM instructions and passes
These only have outlines added to them for now i.e. the names of the passes and
instructions, but no descriptions. These intentionally include passes and
instructions added by both Kitsune and Tapir - the latter because they do not
seem to be documented in this level of detail elsewhere.

Rename existing documents for consistency. The same for the titles of some of
these documents.

Expand the glossary and tweak the text in the FAQ. Also add explicit link
targets to the FAQ.